move all main files in /core
