--- packages/library/src/commandsDefault.ts
@@ -1,8 +1,0 @@
-export const ping: TCommand = async ({ content, response }): Promise<void> => {
-  const dateMessage = content.message().createdTimestamp
-  const dateNow = Date.now()
-
-  const ping = dateNow - dateMessage
-
-  response.general(`🛰️ Pong - duración: **${ping} ms**`)
-}
