+++ packages/library/src/index.ts
@@ -1,2 +1,2 @@
 export { TOptions, FLAGS, TCommand, TMessage, TChannels } from '~ENTITIES/types'
-export { Bot } from './Bot'
+export { Bot } from '~CORE/Bot'
